requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
pdfplumber==0.10.3
PyPDF2
firebase-admin==6.4.0
pyrebase4==4.7.1
python-dotenv==1.0.1
kivy==2.3.0
kivymd==1.1.1
buildozer==1.5.0
pillow==10.2.0
numpy==1.26.3
python-dateutil==2.8.2
pytz==2024.1
requests<2.30.0
urllib3<2.0.0
cryptography==42.0.2
pymupdf==1.18.19